Rename all packages etc. to reflect additional functionaliity
