+++ src/main/java/com/viviquity/jenkins/packageparameter/PersistentPackageParameterDefinition.java
@@ -1,4 +1,4 @@
-package com.viviquity.jenkins.yumparameter;
+package com.viviquity.jenkins.packageparameter;
 
 import hudson.Extension;
 import hudson.model.ParameterValue;
@@ -19,7 +19,7 @@ import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.export.Exported;
 
-import com.viviquity.jenkins.yumparameter.aws.AwsClientReader;
+import com.viviquity.jenkins.packageparameter.aws.AwsClientReader;
 
 /**
  * yum-parameter
