+++ src/main/java/com/viviquity/jenkins/packageparameter/aws/apt/PackagesFormatter.java
@@ -1,11 +1,11 @@
-package com.viviquity.jenkins.yumparameter.aws.apt;
+package com.viviquity.jenkins.packageparameter.aws.apt;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
-import com.viviquity.jenkins.yumparameter.aws.apt.model.Package;
+import com.viviquity.jenkins.packageparameter.aws.apt.model.Package;
 
 public class PackagesFormatter {
 	private VersionComparator versionComparator = new VersionComparator();
