+++ src/main/java/com/viviquity/jenkins/packageparameter/aws/yum/model/package-info.java
@@ -5,4 +5,4 @@
  * @since 16/02/15
  */
 @javax.xml.bind.annotation.XmlSchema(namespace = "http://linux.duke.edu/metadata/common", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
-package com.viviquity.jenkins.yumparameter.aws.yum.model;
+package com.viviquity.jenkins.packageparameter.aws.yum.model;
