+++ src/main/resources/com/viviquity/jenkins/packageparameter/PersistentPackageParameterDefinition/help-bucketName.html
